01/22 :sparkles: : 리다이렉션 로직 추가
